Fix scrolling on mobile using items as handle
DragDropTouch.js would prevent a user on mobile from using any
screen space covered by the items as a handle for scrolling.   When
lists got large, this could cover almost the entire screen, effectively
preventing all scrolling.  The solution was to remove a premature call
to preventDefault(), which happened before the timeout that
determined whether the touch was a press-and-hold (to drag).  With
the default behavior enabled, I had to balance that by removing the
dispatch("click"), so there wouldn't be two click events for every tap.